worker/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stream-movies-worker"
version = "0.1.0"
description = "Celery worker scaffold for media jobs"
requires-python = ">=3.11"
dependencies = [
  "azure-identity>=1.24,<2.0",
  "azure-storage-blob>=12.27,<13.0",
  "celery[redis]>=5.4,<6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
